+++ sdk/keyvault/keyvault-common/review/keyvault-common.api.md
@@ -1,0 +1,24 @@
+## API Report File for "@azure/keyvault-common"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+// @public (undocumented)
+export function assertThrowsAbortError(cb: () => Promise<any>): Promise<void>;
+
+// @public (undocumented)
+export function retry<T>(target: () => Promise<T>, delay?: number, timeout?: number, increaseFactor?: number): Promise<T>;
+
+// @public (undocumented)
+export const testPollerProperties: {
+    intervalInMs: number | undefined;
+};
+
+// @public (undocumented)
+export function uniqueString(): string;
+
+
+// (No @packageDocumentation comment for this package)
+
+```
