+++ sdk/keyvault/keyvault-common/src/test/exceptions.ts
@@ -1,0 +1,16 @@
+import * as assert from "assert";
+
+export async function assertThrowsAbortError(cb: () => Promise<any>) {
+  let passed = false;
+  try {
+    await cb();
+    passed = true;
+  } catch (e) {
+    console.log(`name: ${e.name}, message: ${e.message}`);
+    assert.equal(e.name, "AbortError");
+    assert.equal(e.message, "The operation was aborted.");
+  }
+  if (passed) {
+    throw new Error("Expected cb to throw an AbortError");
+  }
+}
