+++ sdk/keyvault/keyvault-common/src/test/index.ts
@@ -1,1 +1,3 @@
+import "./polyfills"
+export * from "./exceptions";
 export * from "./recorderUtils";
