+++ sdk/keyvault/keyvault-common/src/test/polyfills.ts
@@ -1,0 +1,4 @@
+// Async iterator's polyfill for Node 8
+if (!Symbol || !(Symbol as any).asyncIterator) {
+  (Symbol as any).asyncIterator = Symbol.for("Symbol.asyncIterator");
+}
