+++ sdk/keyvault/keyvault-keys/test/list.test.ts
@@ -3,12 +3,10 @@
 
 import * as assert from "assert";
 import { KeyClient } from "../src";
-import { testPollerProperties } from "./utils/recorderUtils";
-import { retry } from "./utils/recorderUtils";
+import { retry, testPollerProperties, assertThrowsAbortError } from "@azure/keyvault-common";
 import { env, Recorder } from "@azure/test-utils-recorder";
 import { authenticate } from "./utils/testAuthentication";
 import TestClient from "./utils/testClient";
-import { assertThrowsAbortError } from "./utils/utils.common";
 
 describe("Keys client - list keys in various ways", () => {
   const keyPrefix = `recover${env.KEY_NAME || "KeyName"}`;
