+++ sdk/keyvault/keyvault-keys/test/lro.delete.test.ts
@@ -3,7 +3,7 @@
 
 import * as assert from "assert";
 import { KeyClient, DeletedKey } from "../src";
-import { testPollerProperties } from "./utils/recorderUtils";
+import { testPollerProperties } from "@azure/keyvault-common";
 import { env, Recorder } from "@azure/test-utils-recorder";
 import { authenticate } from "./utils/testAuthentication";
 import TestClient from "./utils/testClient";
